split server.go to server.go, handlers.go and routes.go
